+++ eslint.config.js
@@ -50,6 +50,14 @@ export default tseslint.config(
           allowNumber: true,
         },
       ],
+      "wc/guard-super-call": "off", // Disabled as it's checked by TypeScript
+      "lit-a11y/anchor-is-valid": [
+        "error",
+        {
+          allowHash: false,
+          aspects: ["invalidHref", "preferButton"],
+        },
+      ],
     },
   },
   {