+++ src/components/progress-indicator-item/index.ts
@@ -1,0 +1,1 @@
+export * from "./daikin-progress-indicator-item";